perbaikan database dan tampilan untuk form audit sementara

+++ database/migrations/2026_04_01_071515_create_audit_responses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('audit_id')->constrained()->cascadeOnDelete();
             $table->foreignId('cobit_question_id')->constrained()->cascadeOnDelete();
-            $table->integer('score'); 
+            $table->decimal('score', 3, 1); 
             $table->string('evidence_file')->nullable(); 
             $table->text('notes')->nullable();
             $table->timestamps();
